+++ R/model_config.R
@@ -1,49 +1,49 @@
 #' Internal function for global options
-#' @importFrom GlobalOptions setGlobalOptions
+#' @importFrom GlobalOptions set_opt .v
 #' @keywords internal
 #' @seealso \code{\link{modelselect_opt}}
-lt <- setGlobalOptions(
+lt <- set_opt(
   MPIcalc = list(.value = FALSE),
   nbclust = list(.value = function() parallel::detectCores()),
   graph = list(.value = FALSE),
   datatype = list(.value = "ContPosNull"),
   modeltypes = list(.value = function()
-    if (get_opt_value('datatype') == "PA") {
+    if (.v$datatype == "PA") {
       c("PAGLM", "PAGLMns", "PA")
-    } else if (get_opt_value('datatype') == "Cont") {
+    } else if (.v$datatype == "Cont") {
       c("ContGLM", "ContGLMns", "Cont")
-    } else if (get_opt_value('datatype') == "ContPosNull") {
+    } else if (.v$datatype == "ContPosNull") {
       c("LogContGLM", "GammaGLM", "ContGLM",
         "LogContGLMns", "GammaGLMns", "ContGLMns",
         "LogCont", "Gamma", "TweedGLM", "TweedGLMns",
         "Cont")
-    } else if (get_opt_value('datatype') == "Count") {
+    } else if (.v$datatype == "Count") {
       c("CountGLM", "CountGLMns", "Count")
-    } else if (get_opt_value('datatype') == "TweedGLM") {
+    } else if (.v$datatype == "TweedGLM") {
       c("TweedGLMns", "TweedGLM")
-    } else if (get_opt_value('datatype') == "PosCount") {
+    } else if (.v$datatype == "PosCount") {
       c("PosCountGLM", "PosCountGLMns", "PosCount")
-    } else if (get_opt_value('datatype') == "KrigeGLM") {
+    } else if (.v$datatype == "KrigeGLM") {
       c("KrigeGLM")
-    } else if (get_opt_value('datatype') == "KrigeGLM.dist") {
+    } else if (.v$datatype == "KrigeGLM.dist") {
       c("TweedGLM", "KrigeGLM.dist.lambda", "KrigeGLM.dist")}),
   MaxDist = list(.value = function()
-    ifelse(grepl("KrigeGLM", get_opt_value('datatype')), 6000, NA)),
+    ifelse(grepl("KrigeGLM", .v$datatype), 6000, NA)),
   Phi = list(.value = function()
-    ifelse(grepl("KrigeGLM", get_opt_value('datatype')), c(1000, get_opt_value('MaxDist')), NA)),
+    ifelse(grepl("KrigeGLM", .v$datatype), c(1000, .v$MaxDist), NA)),
   Model = list(.value = function()
-    if (get_opt_value('datatype') == "KrigeGLM") {"exponential"
-    } else if (get_opt_value('datatype') == "KrigeGLM.dist") {
+    if (.v$datatype == "KrigeGLM") {"exponential"
+    } else if (.v$datatype == "KrigeGLM.dist") {
       c(NA, "exponential", "exponential")
-    } else {rep(NA, length(get_opt_value('modeltypes')))}),
-    Lambda = list(.value = function() if (get_opt_value('datatype') == "KrigeGLM") {1
-    } else if (get_opt_value('datatype') == "KrigeGLM.dist") {
+    } else {rep(NA, length(.v$modeltypes))}),
+    Lambda = list(.value = function() if (.v$datatype == "KrigeGLM") {1
+    } else if (.v$datatype == "KrigeGLM.dist") {
       c(NA, -1.5, 1)
-    } else {rep(NA, length(get_opt_value('modeltypes')))}),
-    fix.Lambda = list(.value = function() if (get_opt_value('datatype') == "KrigeGLM") {1
-    } else if (get_opt_value('datatype') == "KrigeGLM.dist") {
+    } else {rep(NA, length(.v$modeltypes))}),
+    fix.Lambda = list(.value = function() if (.v$datatype == "KrigeGLM") {1
+    } else if (.v$datatype == "KrigeGLM.dist") {
       c(TRUE, FALSE, TRUE)
-    } else {rep(TRUE, length(get_opt_value('modeltypes')))}),
+    } else {rep(TRUE, length(.v$modeltypes))}),
   lcc_proj = list(.value = "+init=epsg:2154"),
   Max_nb_Var = list(.value = 7),
   Max_K = list(.value = 5),
@@ -51,29 +51,24 @@ lt <- setGlobalOptions(
   Max_K_Poly = list(.value = 4),
   fixXI = list(.value = 0),
   Interaction = list(.value = function()
-    ifelse(grepl("KrigeGLM", get_opt_value('datatype')), FALSE, FALSE)),
+    ifelse(grepl("KrigeGLM", .v$datatype), FALSE, FALSE)),
   MinNbModel = list(.value = 4),
   k_fold = list(.value = 5),
   N_k_fold = list(.value = 20),
-  nbMC = list(.value = function() get_opt_value('N_k_fold')*get_opt_value('k_fold')),
+  nbMC = list(.value = function() .v$N_k_fold*.v$k_fold),
   seqthd = list(.value = seq(0.1, 0.9, 0.01)),
   lim_pvalue = list(.value = 0.001),
   lim_pvalue_final = list(.value = 0.005),
-  Y.max = list(.value = function() ifelse(grepl("PA", get_opt_value('datatype')), 1, 2)),
-  seed = list(.value = 20),
-  get_opt_value_fun = TRUE
+  Y.max = list(.value = function() ifelse(grepl("PA", .v$datatype), 1, 2)),
+  seed = list(.value = 20)
 )
-
-#' Internal function for global options
-#' @keywords internal
-#' @seealso \code{\link{modelselect_opt}}
-get_opt_value <- lt$get_opt_value
 
 #' Modelselect configuration file
 #'
 #' @param RESET reset options to default
-#' @param LOCAL set to true to use modified options in a local environment but not in global.
-#' @param READ.ONLY Logical To not modify the options
+#' @param LOCAL Logical. set to true to use modified options in a local environment but not in global.
+#' @param READ.ONLY Logical. To not modify the options
+#' @param ADD Logical. New options can be added after the option function is created by explicitely specifying ADD = TRUE
 #' @param ... Other options as listed below
 #'
 #' @return A list with all options
@@ -167,6 +162,9 @@ get_opt_value <- lt$get_opt_value
 #' then GAM is fitted with library mgcv.
 #'
 #' }
+#'
+#' @importFrom GlobalOptions set_opt .v
+#'
 #' @export
 #'
 #' @examples
@@ -183,4 +181,63 @@ get_opt_value <- lt$get_opt_value
 #' modelselect_opt("modeltypes") <- c("PA", "PAGLM")
 #' }
 
-modelselect_opt <- lt$opt_fun
+modelselect_opt <- set_opt(
+  MPIcalc = list(.value = FALSE),
+  nbclust = list(.value = function() parallel::detectCores()),
+  graph = list(.value = FALSE),
+  datatype = list(.value = "ContPosNull"),
+  modeltypes = list(.value = function()
+    if (.v$datatype == "PA") {
+      c("PAGLM", "PAGLMns", "PA")
+    } else if (.v$datatype == "Cont") {
+      c("ContGLM", "ContGLMns", "Cont")
+    } else if (.v$datatype == "ContPosNull") {
+      c("LogContGLM", "GammaGLM", "ContGLM",
+        "LogContGLMns", "GammaGLMns", "ContGLMns",
+        "LogCont", "Gamma", "TweedGLM", "TweedGLMns",
+        "Cont")
+    } else if (.v$datatype == "Count") {
+      c("CountGLM", "CountGLMns", "Count")
+    } else if (.v$datatype == "TweedGLM") {
+      c("TweedGLMns", "TweedGLM")
+    } else if (.v$datatype == "PosCount") {
+      c("PosCountGLM", "PosCountGLMns", "PosCount")
+    } else if (.v$datatype == "KrigeGLM") {
+      c("KrigeGLM")
+    } else if (.v$datatype == "KrigeGLM.dist") {
+      c("TweedGLM", "KrigeGLM.dist.lambda", "KrigeGLM.dist")}),
+  MaxDist = list(.value = function()
+    ifelse(grepl("KrigeGLM", .v$datatype), 6000, NA)),
+  Phi = list(.value = function()
+    ifelse(grepl("KrigeGLM", .v$datatype), c(1000, .v$MaxDist), NA)),
+  Model = list(.value = function()
+    if (.v$datatype == "KrigeGLM") {"exponential"
+    } else if (.v$datatype == "KrigeGLM.dist") {
+      c(NA, "exponential", "exponential")
+    } else {rep(NA, length(.v$modeltypes))}),
+  Lambda = list(.value = function() if (.v$datatype == "KrigeGLM") {1
+  } else if (.v$datatype == "KrigeGLM.dist") {
+    c(NA, -1.5, 1)
+  } else {rep(NA, length(.v$modeltypes))}),
+  fix.Lambda = list(.value = function() if (.v$datatype == "KrigeGLM") {1
+  } else if (.v$datatype == "KrigeGLM.dist") {
+    c(TRUE, FALSE, TRUE)
+  } else {rep(TRUE, length(.v$modeltypes))}),
+  lcc_proj = list(.value = "+init=epsg:2154"),
+  Max_nb_Var = list(.value = 7),
+  Max_K = list(.value = 5),
+  Max_K_te = list(.value = 3),
+  Max_K_Poly = list(.value = 4),
+  fixXI = list(.value = 0),
+  Interaction = list(.value = function()
+    ifelse(grepl("KrigeGLM", .v$datatype), FALSE, FALSE)),
+  MinNbModel = list(.value = 4),
+  k_fold = list(.value = 5),
+  N_k_fold = list(.value = 20),
+  nbMC = list(.value = function() .v$N_k_fold*.v$k_fold),
+  seqthd = list(.value = seq(0.1, 0.9, 0.01)),
+  lim_pvalue = list(.value = 0.001),
+  lim_pvalue_final = list(.value = 0.005),
+  Y.max = list(.value = function() ifelse(grepl("PA", .v$datatype), 1, 2)),
+  seed = list(.value = 20)
+)